Simulating 2 splits

Simulating 2 splits

split2(
  data,
  est.lambda = FALSE,
  lambda.val = 0,
  n.exposure,
  n.outcome,
  n.overlap,
  cor.xy,
  pi = 0.5,
  mr_method = "mr_ivw",
  threshold = 5e-08
)

Arguments

data

A data frame to be inputted by the user containing summary statistics from the exposure and outcome GWASs. It must have at least five columns with column names SNP, beta.exposure, beta.outcome, se.exposure, and se.outcome. Each row must correspond to a unique SNP, identified by SNP.

est.lambda

A logical value.

lambda.val

A numerical value.

n.exposure

A numerical value to be specified by the user which is equal to the number of individuals that were in the exposure GWAS.

n.outcome

A numerical value to be specified by the user which is equal to the number of individuals that were in the outcome GWAS.

n.overlap

A numerical value to be specified by the user which is equal to the number of individuals that were in both the exposure and outcome GWAS.

cor.xy

A numerical value to be specified by the user which is equal to the observed correlation between the exposure and the outcome. This value must be between -1 and 1.

pi

A numerical value. The default setting is pi=0.5.

mr_method

A string. The default setting is mr_method="mr_ivw".

threshold

A numerical value. The default setting is threshold=5e-8.

Value

A data frame