Simulating 2 splits with MR-SimSS

split2 is a function which is used by the main function, mr_simss in order to perform the 2 split approach of the method, MR-SimSS.

split2(
  data,
  est.lambda = FALSE,
  lambda.val = 0,
  n.exposure = 1,
  n.outcome = 1,
  n.overlap = 1,
  cor.xy = 0,
  pi = 0.5,
  mr_method = "mr_ivw",
  threshold = 5e-08
)

Arguments

data

A data frame to be inputted by the user containing summary statistics from the exposure and outcome GWASs. It must have at least five columns with column names SNP, beta.exposure, beta.outcome, se.exposure and se.outcome. Each row must correspond to a unique SNP, identified by SNP.

est.lambda

A logical which allows the user to specify if they have used the function, est_lambda, to obtain an estimate for lambda, a term used to describe the correlation between the outcome and exposure effect sizes. This correlation is affected by the number of overlapping samples between the two GWASs and the correlation between the exposure and the outcome. Thus, it is recommended to use est_lambda if the fraction of overlap and the correlation between exposure and outcome are unknown. The default setting is est.lambda=FALSE.

lambda.val

A numerical value which should be specified by the user if est.lambda=TRUE. It should be equal to the value returned from using the function est_lambda. The default setting is lambda.val=0.

n.exposure

A numerical value to be specified by the user which is equal to the number of individuals that were in the exposure GWAS. It should be specified by the user if est.lambda=FALSE. The default setting is n.exposure=1.

n.outcome

A numerical value to be specified by the user which is equal to the number of individuals that were in the outcome GWAS. It should be specified by the user if est.lambda=FALSE. The default setting is n.outcome=1.

n.overlap

A numerical value to be specified by the user which is equal to the number of individuals that were in both the exposure and outcome GWAS. It should be specified by the user if est.lambda=FALSE. The default setting is n.overlap=1.

cor.xy

A numerical value to be specified by the user which is equal to the observed correlation between the exposure and the outcome. It should be specified by the user if est.lambda=FALSE. The default setting is cor.xy=0. If this value is unknown, the user is encouraged to use the function est_lambda.

pi

A numerical value which determines the fraction of the first split. This is the fraction that will be used for SNP selection. The default setting is pi=0.5.

mr_method

A string which specifies the MR method that MR-SimSS works in combination with. It is possible to use any method outputted in the list TwoSampleMR::mr_method_list()$obj. However, it is currently advised that the user chooses "mr_ivw" or "mr_raps". The default setting is mr_method="mr_ivw".

threshold

A numerical value which specifies the threshold used to select instrument SNPs for MR at each iteration. The default setting is threshold=5e-8.

Value

A data frame which contains the output from one iteration. It is in a similar style as the output from using the function mr from the TwoSampleMR R package. The MR method used, the number of instrument SNPs, the causal effect estimate, it associated standard error and p-value are all outputted.